# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Prayer Selector API",
    "description": "This API allows for the user to get, update, add, and delete data from the Prayer Selector database",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/about": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api-docs/": {
      "get": {
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/classes/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "class",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "department": {
                  "type": "string",
                  "example": ""
                },
                "code": {
                  "type": "string",
                  "example": ""
                },
                "name": {
                  "type": "string",
                  "example": ""
                },
                "section": {},
                "capacity": {},
                "students": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "fname": {
                        "type": "string",
                        "example": ""
                      },
                      "lname": {
                        "type": "string",
                        "example": ""
                      },
                      "profilepic": {}
                    },
                    "required": [
                      "fname",
                      "lname",
                      "profilepic"
                    ]
                  }
                }
              },
              "required": [
                "department",
                "code",
                "name",
                "section",
                "capacity",
                "students"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/classes/{_id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "class",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "department": {
                  "type": "string",
                  "example": ""
                },
                "code": {
                  "type": "string",
                  "example": ""
                },
                "name": {
                  "type": "string",
                  "example": ""
                },
                "section": {},
                "capacity": {},
                "students": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "fname": {
                        "type": "string",
                        "example": ""
                      },
                      "lname": {
                        "type": "string",
                        "example": ""
                      },
                      "profilepic": {}
                    },
                    "required": [
                      "fname",
                      "lname",
                      "profilepic"
                    ]
                  }
                }
              },
              "required": [
                "department",
                "code",
                "name",
                "section",
                "capacity",
                "students"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}